fix(tools): stop the rsync remote end claiming a grouping the API never states
Round 1's review found one MEDIUM and four LOWs, all confirmed against the
pinned client's declarations before being acted on.

MEDIUM: the description said `mode` decides whether `remote_host` and
`remote_port` or `remote_module` describe the far end. Nothing in
`RsyncTaskEntry` ties a field to a mode, and an SSH task's hostname and port
can instead live inside the `SSHCredentials` this tool deliberately does not
read — so a null `remote_host` beside a named credential is not the absence
the sentence promised. Each field is now described as what the task itself
records, and the description says the credential can hold the remote host and
that nothing in this catalog describes a stored SSH credential further.

LOW, all four: `timeout_seconds` carried a unit the pinned surface never
states, which decision #96 exists to stop — it is `timeout` now, with the
description saying no unit is reported. `tasks_recent_runs` claimed a cron job
never appears there "however often it has run", but `cronjob.run` is a job
method, so a manual run does; the claim is narrowed to scheduled runs. The
tests were moved back into `tasks.spec.ts`: the merged file is about 1,311
lines, so #87's 1,500-line split exception was never met and the recorded
justification for the new file was a number anyone could check. The rsync
fixture no longer carries an SSH-mode row with both a credential and its own
remote host, a combination the middleware does not send.

The two CLAUDE.md entries move with the code.
